requirement.txt
astroid==2.3.3
bson==0.5.8
colorama==0.4.3
cycler==0.10.0
dataclasses==0.7
Django==2.1.3
django-sslserver==0.22
djangorestframework==3.11.0
djongo==1.3.0
dnspython==1.16.0
isort==4.3.21
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
matplotlib==3.0.2
mccabe==0.6.1
mysqlclient==1.3.13
numpy==1.15.4
Pillow==7.0.0
psycopg2==2.7.6.1
Pygments==2.5.2
pylint==2.4.4
pymongo==3.10.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
six==1.14.0
sqlparse==0.2.4
typed-ast==1.4.1
wrapt==1.11.2
